+++ CHANGELOG.md
@@ -5,11 +5,29 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 
-## 2020-02-26
+## 2020-03-03
 
 ### Added
 
 - World bank tariff (data-store-service) pipeline
+
+
+## 2020-02-27
+
+### Changed
+
+- Add created by id to interactions dataset
+
+
+## 2020-02-26
+
+### Changed
+
+- Update monthly static FDI report query
+  - Join string arrays with semicolons
+  - Do not include empty string for `other_business_activities` field 
+- Ensure omis orders are included to the end of the month for the given run date
+- Set null and false values to false in export wins early cut
 
 
 ## 2020-02-25